+++ packages/generator/generator-common/src/newAction/index.ts
@@ -1,4 +1,3 @@
 export * from './common';
 export * from './mwa';
 export * from './module';
-export * from './monorepo';