+++ packages/generator/generators/module-test-generator/src/locale/zh.ts
@@ -1,3 +1,0 @@
-export const ZH_LOCALE = {
-  success: `添加完成后，你可运行 {packageManager} test 运行测试用例。`,
-};