+++ packages/cli/plugin-lint/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/plugin-lint
 
+## 2.52.0
+
+### Patch Changes
+
+- @modern-js/tsconfig@2.52.0
+- @modern-js/utils@2.52.0
+
 ## 2.51.0
 
 ### Patch Changes