+++ packages/solutions/module-tools/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/module-tools
 
+## 2.52.0
+
+### Patch Changes
+
+- @modern-js/core@2.52.0
+- @modern-js/plugin-changeset@2.52.0
+- @modern-js/plugin-i18n@2.52.0
+- @modern-js/plugin-lint@2.52.0
+- @modern-js/plugin@2.52.0
+- @modern-js/types@2.52.0
+- @modern-js/utils@2.52.0
+
 ## 2.51.0
 
 ### Patch Changes